change GroverOptimizerResult to have samples
