+++ qiskit_optimization/algorithms/optimization_algorithm.py
@@ -20,6 +20,7 @@ from warnings import warn
 
 import numpy as np
 
+from qiskit.opflow import StateFn, DictStateFn
 from .. import QiskitOptimizationError
 from ..converters.quadratic_program_to_qubo import (QuadraticProgramToQubo,
                                                     QuadraticProgramConverter)
@@ -421,3 +422,94 @@ class OptimizationAlgorithm(ABC):
                             variables=problem.variables,
                             status=cls._get_feasibility_status(problem, x),
                             **kwargs)
+
+    def _interpret_samples(self, problem: QuadraticProgram, raw_samples: List[SolutionSample]) \
+            -> List[SolutionSample]:
+        prob = {}  # type: dict
+        array = {}
+        for sample in raw_samples:
+            x = sample.x
+            for converter in self._converters[::-1]:
+                x = converter.interpret(x)
+            key = tuple(x)
+            prob[key] = prob.get(key, 0.0) + sample.probability
+            array[key] = x
+
+        samples = []
+        for key, x in array.items():
+            probability = prob[key]
+            fval = problem.objective.evaluate(x)
+            status = self._get_feasibility_status(problem, x)
+            samples.append(SolutionSample(x, fval, probability, status))
+
+        return sorted(samples,
+                      key=lambda v: (v.status.value, problem.objective.sense.value * v.fval))
+
+    def _eigenvector_to_solutions(self, eigenvector: Union[dict, np.ndarray, StateFn],
+                                  qubo: QuadraticProgram,
+                                  min_probability: float = 1e-6,
+                                  ) -> List[SolutionSample]:
+        """Convert the eigenvector to the bitstrings and corresponding eigenvalues.
+
+        Args:
+            eigenvector: The eigenvector from which the solution states are extracted.
+            qubo: The QUBO to evaluate at the bitstring.
+            min_probability: Only consider states where the amplitude exceeds this threshold.
+
+        Returns:
+            For each computational basis state contained in the eigenvector, return the basis
+            state as bitstring along with the QUBO evaluated at that bitstring and the
+            probability of sampling this bitstring from the eigenvector.
+
+        Examples:
+            >>> op = MatrixOp(numpy.array([[1, 1], [1, -1]]) / numpy.sqrt(2))
+            >>> eigenvectors = {'0': 12, '1': 1}
+            >>> print(eigenvector_to_solutions(eigenvectors, op))
+            [('0', 0.7071067811865475, 0.9230769230769231),
+            ('1', -0.7071067811865475, 0.07692307692307693)]
+
+            >>> op = MatrixOp(numpy.array([[1, 1], [1, -1]]) / numpy.sqrt(2))
+            >>> eigenvectors = numpy.array([1, 1] / numpy.sqrt(2), dtype=complex)
+            >>> print(eigenvector_to_solutions(eigenvectors, op))
+            [('0', 0.7071067811865475, 0.4999999999999999),
+            ('1', -0.7071067811865475, 0.4999999999999999)]
+
+        Raises:
+            TypeError: If the type of eigenvector is not supported.
+        """
+        if isinstance(eigenvector, DictStateFn):
+            eigenvector = {bitstr: val ** 2 for (bitstr, val) in eigenvector.primitive.items()}
+        elif isinstance(eigenvector, StateFn):
+            eigenvector = eigenvector.to_matrix()
+
+        def generate_solution(bitstr, qubo, probability):
+            x = np.fromiter(list(bitstr), dtype=int)
+            fval = qubo.objective.evaluate(x)
+            return SolutionSample(x=x, fval=fval, probability=probability,
+                                  status=OptimizationResultStatus.SUCCESS)
+
+        solutions = []
+        if isinstance(eigenvector, dict):
+            all_counts = sum(eigenvector.values())
+            # iterate over all samples
+            for bitstr, count in eigenvector.items():
+                sampling_probability = count / all_counts
+                # add the bitstring, if the sampling probability exceeds the threshold
+                if sampling_probability >= min_probability:
+                    solutions.append(generate_solution(bitstr, qubo, sampling_probability))
+
+        elif isinstance(eigenvector, np.ndarray):
+            num_qubits = int(np.log2(eigenvector.size))
+            probabilities = np.abs(eigenvector * eigenvector.conj())
+
+            # iterate over all states and their sampling probabilities
+            for i, sampling_probability in enumerate(probabilities):
+                # add the i-th state if the sampling probability exceeds the threshold
+                if sampling_probability >= min_probability:
+                    bitstr = '{:b}'.format(i).rjust(num_qubits, '0')[::-1]
+                    solutions.append(generate_solution(bitstr, qubo, sampling_probability))
+
+        else:
+            raise TypeError('Unsupported format of eigenvector. Provide a dict or numpy.ndarray.')
+
+        return solutions
